added some tests, refactored callbacks and datasets
