+++ openml_pytorch/custom_datasets/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module contains the custom datasets for OpenML datasets.
+"""
+from .image_dataset import OpenMLImageDataset
+from .tabular_dataset import OpenMLTabularDataset
+
+__all__ = ["OpenMLTabularDataset", "OpenMLImageDataset"]
